Refactor character handling for better extensability with new systems
