+++ src/styles/styles.tsx
@@ -1,4 +1,5 @@
 import { useTheme, Palette, PaletteColor, alpha, Paper } from "@mui/material";
+import { Theme } from "@mui/material/styles";
 
 type KeysByType<O, T> = {
   [K in keyof O]-?: T extends O[K] ? K : never;
@@ -6,42 +7,36 @@ type KeysByType<O, T> = {
 
 export type Colors = KeysByType<Palette, PaletteColor>;
 
-export function MenuStyle() {
-  const theme = useTheme();
+export const menuStyle = (theme: Theme) => ({
+  "& .MuiButtonBase-root.MuiMenuItem-root:hover": {
+    backgroundColor: alpha(theme.palette.primary.main, 0.12),
+    textShadow: `0 0 6px ${theme.palette.primary.main}`,
+    transition: "all 0.2s ease-in-out",
+    color: theme.palette.primary.main,
+  },
+  "& .MuiButtonBase-root.MuiMenuItem-root": {
+    color: theme.palette.text.primary,
+  },
+  "& .MuiList-root.MuiMenu-list": {
+    padding: "8px",
+  },
+  "& .MuiPaper-root.MuiPopover-paper.MuiMenu-paper": {
+    color: theme.palette.primary.main,
+  },
+  "& .MuiPaper-root": {
+    backgroundImage: "none",
+    backgroundColor: theme.palette.background.default,
+    boxShadow: `inset 0px 0px 0px 1px ${
+      theme.palette.primary.dark
+    }, 0 0 8px ${alpha(theme.palette.primary.main, 0.25)}`,
+    borderRadius: "6px",
+    "&::before": {
+      backgroundColor: theme.palette.background.default,
+    },
+  },
+});
 
-  return {
-    "& .MuiButtonBase-root.MuiMenuItem-root:hover": {
-      backgroundColor: alpha(theme.palette.primary.main, 0.12),
-      textShadow: `0 0 6px ${theme.palette.primary.main}`,
-      transition: "all 0.2s ease-in-out",
-      color: theme.palette.primary.main,
-    },
-    "& .MuiButtonBase-root.MuiMenuItem-root": {
-      color: theme.palette.text.primary,
-    },
-    "& .MuiList-root.MuiMenu-list": {
-      padding: "8px",
-    },
-    "& .MuiPaper-root.MuiPopover-paper.MuiMenu-paper": {
-      color: theme.palette.primary.main,
-    },
-    "& .MuiPaper-root": {
-      backgroundImage: "none",
-      backgroundColor: theme.palette.background.default, // Use background color from theme
-      boxShadow: `inset 0px 0px 0px 1px ${
-        theme.palette.primary.dark
-      }, 0 0 8px ${alpha(theme.palette.primary.main, 0.25)}`, // Added outer glow
-      borderRadius: "6px",
-      "&::before": {
-        backgroundColor: theme.palette.background.default, // Use background color from theme
-      },
-    },
-  };
-}
-
-export function AutocompleteStyle() {
-  const theme = useTheme();
-
+export const autocompleteStyle = (theme: Theme) => {
   const baseColor = theme.palette.primary.main;
   const editColor = theme.palette.primary.light;
   const darkColor = theme.palette.primary.dark;
@@ -147,107 +142,82 @@ export function AutocompletePaper({
   );
 }
 
-export function ModalStyle() {
-  const theme = useTheme();
-  return {
-    position: "absolute" as "absolute",
-    top: "50%",
-    left: "50%",
-    transform: "translate(-50%, -50%)",
-    width: 400,
-    bgcolor: theme.palette.background.default, // Use background color from theme
-    border: `1px solid ${theme.palette.primary.dark}`, // Slight border for visibility
-    boxShadow: `0 0 12px ${alpha(theme.palette.primary.main, 0.25)}`, // Soft glow effect
-    borderRadius: "8px",
-    p: 4,
-  };
-}
+export const modalStyle = (theme: Theme) => ({
+  position: "absolute" as const,
+  top: "50%",
+  left: "50%",
+  transform: "translate(-50%, -50%)",
+  width: 400,
+  bgcolor: theme.palette.background.default,
+  border: `1px solid ${theme.palette.primary.dark}`,
+  boxShadow: `0 0 12px ${alpha(theme.palette.primary.main, 0.25)}`,
+  borderRadius: "8px",
+  p: 4,
+});
 
 // ================================
 // CharacterCard & CharacterManager styles
 // ================================
 
 // Action Buttons box style
-export function ActionButtonsBoxStyle() {
-  return {
-    display: "flex",
-    justifyContent: "space-around",
-    width: "100%",
-  };
-}
+export const actionButtonsBoxStyle = {
+  display: "flex",
+  justifyContent: "space-around",
+  width: "100%",
+};
 
 // Individual action button style
-export function ActionButtonStyle() {
-  return {
-    textAlign: "center",
-  };
-}
+export const actionButtonStyle = {
+  textAlign: "center" as const,
+};
 
 // Accordion grid style
-export function AccordionGridStyle() {
-  const theme = useTheme();
-  return {
-    "& .MuiPaper-root": {
-      backgroundImage: "none",
+export const accordionGridStyle = (theme: Theme) => ({
+  "& .MuiPaper-root": {
+    backgroundImage: "none",
+    backgroundColor: theme.palette.background.default,
+    boxShadow: `inset 0px 0px 0px 1px ${theme.palette.primary.dark}, 0px 0px 0px 0px #FF0000, 0px 0px 0px 0px #FF0000`,
+    "&::before": {
       backgroundColor: theme.palette.background.default,
-      boxShadow: `inset 0px 0px 0px 1px ${theme.palette.primary.dark}, 0px 0px 0px 0px #FF0000, 0px 0px 0px 0px #FF0000`,
-      "&::before": {
-        backgroundColor: theme.palette.background.default,
-      },
     },
-  };
-}
+  },
+});
 
 // Card box style
-export function CardBoxStyle() {
-  const theme = useTheme();
-  return {
-    p: 2,
-    border: "0px solid",
-    borderColor: theme.palette.primary.dark,
-    borderRadius: "4px",
-  };
-}
+export const cardBoxStyle = (theme: Theme) => ({
+  p: 2,
+  border: "0px solid",
+  borderColor: theme.palette.primary.dark,
+  borderRadius: "4px",
+});
 
 // Info box and info inner box
-export function InfoBoxStyle() {
-  return { my: 2 };
-}
-
-export function InfoInnerBoxStyle() {
-  return { textAlign: "right" };
-}
+export const infoBoxStyle = { my: 2 };
+export const infoInnerBoxStyle = { textAlign: "right" as const };
 
 // Cyberware & skills
-export function CyberwareBoxStyle() {
-  return {
-    my: 2,
-    display: "flex",
-    flexDirection: "column",
-    alignItems: "flex-start",
-  };
-}
-
-export const SkillsBoxStyle = CyberwareBoxStyle; // Alias; if future difference, split
+export const cyberwareBoxStyle = {
+  my: 2,
+  display: "flex",
+  flexDirection: "column" as const,
+  alignItems: "flex-start",
+};
+export const skillsBoxStyle = cyberwareBoxStyle;
 
 // Damage grid and component box style
-export function trackGridStyle() {
-  return {
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-    flexWrap: "wrap",
-  };
-}
+export const trackGridStyle = {
+  flexDirection: "row" as const,
+  alignItems: "center",
+  justifyContent: "center",
+  flexWrap: "wrap" as const,
+};
 
-export function trackMeterBoxStyle() {
-  return {
-    display: "flex",
-    flexDirection: "row",
-    alignItems: "center",
-    justifyContent: "center",
-  };
-}
+export const trackMeterBoxStyle = {
+  display: "flex",
+  flexDirection: "row" as const,
+  alignItems: "center",
+  justifyContent: "center",
+};
 
 // ================================
 // Damage color helpers
@@ -298,8 +268,7 @@ export function getTrackColor(
 }
 
 // Used for TextField showing current damage percentage
-export function createTrackInputStyle(damagePercentage: number) {
-  const theme = useTheme();
+export const trackInputStyle = (damagePercentage: number) => (theme: Theme) => {
   const damageColor = getTrackColor(damagePercentage, theme.trackColors);
 
   return {
@@ -346,24 +315,23 @@ export function createTrackInputStyle(damagePercentage: number) {
       borderBottomRightRadius: "0.25rem",
     },
   };
-}
+};
 
-export function TextfieldStyle({ color }: { color: Colors }) {
-  const theme = useTheme();
+export const textfieldStyle = (color: Colors) => (theme: Theme) => {
   const baseColor = theme.palette[color].main;
   const editColor = theme.palette[color].light;
   const darkColor = theme.palette[color].dark;
 
   return {
     display: "flex",
-    width: "100%", // Fields take up full width of their container
+    width: "100%",
     paddingTop: 0,
     "& .MuiInputBase-root": {
       color: editColor,
       transition: "all 0.2s ease-in-out",
     },
     "& .MuiOutlinedInput-notchedOutline": {
-      borderColor: darkColor, // Set default border color based on colorType
+      borderColor: darkColor,
     },
     "& .MuiInputBase-root.Mui-disabled": {
       "& > fieldset": {
@@ -371,12 +339,11 @@ export function TextfieldStyle({ color }: { color: Colors }) {
       },
     },
     "& .MuiInputBase-input.Mui-disabled": {
-      WebkitTextFillColor: baseColor, // Adjust text color when disabled
+      WebkitTextFillColor: baseColor,
     },
-    // set color when inactive
     "& .MuiOutlinedInput-root": {
       "&:hover fieldset": {
-        borderColor: baseColor, // Adjust color on hover
+        borderColor: baseColor,
       },
     },
     ".MuiOutlinedInput-root.Mui-focused": {
@@ -388,4 +355,4 @@ export function TextfieldStyle({ color }: { color: Colors }) {
       },
     },
   };
-}
+};
